+++ custom_components/solarcharger/chargers/tesla_fleet_charger.py
@@ -10,9 +10,13 @@ from homeassistant.helpers.device_registry import DeviceEntry
 from ..const import DOMAIN_TESLA_FLEET
 from .charger_chargeable_base import ChargerChargeableBase
 
+# ----------------------------------------------------------------------------
+# ----------------------------------------------------------------------------
 _LOGGER = logging.getLogger(__name__)
 
 
+# ----------------------------------------------------------------------------
+# ----------------------------------------------------------------------------
 class TeslaFleetCharger(ChargerChargeableBase):
     """Implementation of the Charger class for Tesla Fleet chargers."""