+++ custom_components/solarcharger/chargers/teslemetry_charger.py
@@ -10,9 +10,13 @@ from homeassistant.helpers.device_registry import DeviceEntry
 from ..const import DOMAIN_TESLEMETRY
 from .charger_chargeable_base import ChargerChargeableBase
 
+# ----------------------------------------------------------------------------
+# ----------------------------------------------------------------------------
 _LOGGER = logging.getLogger(__name__)
 
 
+# ----------------------------------------------------------------------------
+# ----------------------------------------------------------------------------
 class TeslemetryCharger(ChargerChargeableBase):
     """Implementation of the Charger class for Teslemetry chargers."""