+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Reminder: Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Installation#how-to-delete-and-re-add-the-solarcharger-integration) if upgrading from v0.9.0 or prior.
 ### Fix
 ### New feature
+- Beta release for [MySkoda](https://github.com/skodaconnect/homeassistant-myskoda) support. Need charger device to set current.
+- Beta release for [BYD vehicle](https://github.com/jkaberg/hass-byd-vehicle) support. Need charger device to set current.
 ### Improvement
 
 ## v0.10.4 2026-08-21